feat: add command to import ban list via link
The command supports links from dpaste.com and pastebin.com. The command takes a link to the ban list as an argument and bans all users in the list. The command retries failed bans up to 3 times and waits for 1 second between retries.
The command sends an embed message to the user with the number of bans that were successful and the number of bans that failed. The command also adds a default reason for the ban if no reason is provided in the ban list.